fead: add price plans on landing page
